// src/api/generated/models/MenuCategory.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Menu category response model
 */
export type MenuCategory = {
    name: string;
    description?: (string | null);
    parent_id?: (string | null);
    display_order?: number;
    icon_url?: (string | null);
    is_active?: boolean;
    id: string;
    business_id: string;
    created_at: string;
    updated_at: string;
};